requirements.txt
jsonschema
requests==2.20.0
requests-oauthlib==1.0.0
requests-toolbelt==0.8.0
pyjwt==1.6.4
mock==2.0.0
dpath==1.4.2
google-auth==1.6.1
google-cloud-core==1.1.0
google-api-core[grpc]==1.15.0
googleapis-common-protos==1.6.0
google-cloud-ndb==0.2.1
google-cloud-datastore==1.10.0
Flask==1.1.1
Werkzeug==0.16.0
webargs==5.5.2
